+++ src/main/java/com/stir/cscu9t4practical1/entries/SwimEntry.java
@@ -6,7 +6,7 @@ public class SwimEntry extends Entry {
 
     private final LOCATION location;
 
-    public SwimEntry(String n, ZonedDateTime zonedDateTime, float dist, LOCATION location) {
+    public SwimEntry(String n, ZonedDateTime zonedDateTime, double dist, LOCATION location) {
         super(n, zonedDateTime, dist);
         this.location = location;
     }
